warnings--
git-svn-id: file:///.data/svn_convert/svn@7565 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ trunk/plugins/icq/icqvarious.cpp
@@ -753,7 +753,7 @@ bool SearchWPRequest::answer(ICQBuffer &b, unsigned short nSubType)
 unsigned short ICQClient::findByUin(unsigned long uin)
 {
     if (getState() != Connected)
-        return ~0U;
+        return (unsigned short)~0U;
     serverRequest(ICQ_SRVxREQ_MORE);
     socket()->writeBuffer()
     << ICQ_SRVxREQ_WP_UIN;
@@ -766,9 +766,9 @@ unsigned short ICQClient::findByUin(unsigned long uin)
 unsigned short ICQClient::findByMail(const QString &_mail)
 {
     if (getState() != Connected)
-        return ~0U;
+        return (unsigned short)~0U;
     QCString mail = getContacts()->fromUnicode(NULL, _mail);
-    
+
     serverRequest(ICQ_SRVxREQ_MORE);
     socket()->writeBuffer()
     << ICQ_SRVxREQ_WP_MAIL;
